application.py: Créer une application fonctionnelle
